Правки красоты для travis'а

+++ include/Investigation.cpp
@@ -1,9 +1,9 @@
 // Copyright 2021 wm8
 
 #include "Investigation.h"
-Investigation::Investigation(ExperimentType t, std::vector<long long>& sizes) {
+Investigation::Investigation(ExperimentType t, std::vector<int64_t>& sizes) {
   Investigation::type = t;
-  int id=0;
+  int id = 0;
   for (long long i=log2(sizes[0]/2); pow(2, i) < 3*sizes[sizes.size()-1]/2; i++)
   {
     id++;
@@ -13,11 +13,10 @@ Investigation::Investigation(ExperimentType t, std::vector<long long>& sizes) {
   }
 }
 void Investigation::print(std::ostream& os) {
-
   os<< "investigation\n\ttravel_variant: " << to_string(type) << std::endl;
-  for(Experiment& experiment: experiments)
+  for ( Experiment& experiment : experiments)
   {
-    os<<"\texperiment:\n";
+    os << "\texperiment:\n";
     experiment.print(os);
   }
 }
